split test_proxy into unit and e2e tests
